features relations

+++ src/drivers/mongo.js
@@ -4,10 +4,10 @@ const Driver = require("./driver");
 
 class MongoDriver extends Driver {
     connect() {
-        const {user = "", password = "", host = "localhost", port = "27017", dbname = "baleDB"} = this.config;
+        const {user = "", password = "", host = "localhost", port = "27017", dbname = "baleDB", reset = true} = this.config;
         const url = this.createConnectionString(user, password, host, port, dbname);
         const connectPromise = new Promise((resolve, reject) => {
-            MongoClient.connect(url, function(err, db) {
+            MongoClient.connect(url, function (err, db) {
                 if (err) {
                     reject(err);
                     return;
@@ -18,7 +18,9 @@ class MongoDriver extends Driver {
         return connectPromise         
         .then(db => {
             this.db = db;
-            db.dropDatabase();
+            if (reset) {
+                db.dropDatabase();
+            }
             return "connected!";
         });
     }
